requirements.txt
defusedxml==0.7.1
et_xmlfile==2.0.0
fonttools==4.59.0
fpdf2==2.8.3
openpyxl==3.1.5
pillow==10.4.0
ttkbootstrap==1.14.1
